Fix crash due to access of deleted RequestData object
:Release Notes:
Fix crash due to access of deleted RequestData object

:Detailed Notes:
An object is moved out of the stl container and iterator to
the moved object is used to do erase operation for the same
element. This leads to undefined behaviour. Don't move the
object and just do the erase operation. Since object is
managed by smart pointer, memory management is taken care.

:Testing Performed:
Built locally and verified.

:QA Notes:

:Issues Addressed:
[WRP-20978] [RDX_CRASH][webos] /usr/sbin/storageaccess in
            _ZN7LSUtils11ClientWatch20notifyClientCanceledEPKc
			(storageaccess + 0xa83c0)

Change-Id: I71fe1eed8d20f1d841d42109035c6b4194d4ee35
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.storageaccess/+/361153
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: RajeshGopu IV <rajeshgopu.iv@lge.com>
Tested-by: <sandeep.arjun@lge.com>